Add unit test and plit out spm unit count adult.
